+++ package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "@agm/core": "^1.0.0-beta.6",
     "@angular/animations": "7.2.13",
-    "@angular/cdk": "7.2.1",
+    "@angular/cdk": "^7.2.1",
     "@angular/common": "7.2.13",
     "@angular/compiler": "7.2.13",
     "@angular/core": "7.2.13",
