+++ src/app/add-event/add-event.component.ts
@@ -14,11 +14,8 @@ import { finalize } from 'rxjs/operators';
   styleUrls: ['./add-event.component.scss']
 })
 export class AddEventComponent implements OnInit {
-
-  
   addEventForm: FormGroup;
   authenticated_user: any;
-
   urls = new Array<string>();
   display_image:any;
   ref: AngularFireStorageReference;
@@ -35,7 +32,8 @@ export class AddEventComponent implements OnInit {
     public eventsService:EventsService,
     private toastr: ToastrService,
     private afStorage: AngularFireStorage
-  ) {
+  ) 
+  {
     this.createForm()
    }
