+++ src/app/blogs/blogs.component.scss
@@ -1,0 +1,12 @@
+section{
+    max-width: 700px;
+    margin: 2em auto;
+}
+
+mat-form-field{
+    width: 100%;
+}
+
+img{
+    width: 100%;
+}
