+++ src/app/dashboard/dashboard.component.html
@@ -97,23 +97,17 @@
           </li>
 
           <li>
-            <a [routerLink] = "['/dashboard/imagesvideos']"><i class="image icon"></i>
-              <span>Media</span>
+            <a [routerLink] = "['/dashboard/timeline']"><i class="image icon"></i>
+              <span>Timeline</span>
             </a>
           </li>
           
-          <li id="menu-academico">
+          <li >
             <a [routerLink] = "['/dashboard/reports']"><i class="paste icon"></i>
               <span>Reports</span>
             </a>
           </li>
-          
-          <li id="menu-academico">
-            <a [routerLink] = "['/dashboard/reports']"><i class="lnr lnr-layers"></i>
-              <span>Reports</span>
-            </a>
-          </li>
-            <!-- </li> -->
+            
             <!--for now, the add section option will redirect you to overview-->
           <li id="menu-comunicacao">
             <a [routerLink] = "['/dashboard/AddSection']"><i class="fa fa-smile-o"></i>
