--- src/app/images-videos/images-videos.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit,  ViewChild} from '@angular/core';
-
-
-
-@Component({
-  selector: 'app-images-videos',
-  templateUrl: './images-videos.component.html',
-  styleUrls: ['./images-videos.component.scss']
-})
-
-
-export class ImagesVideosComponent  {
-  
-  
-}
