+++ src/app/mentor/mentor.component.html
@@ -74,7 +74,7 @@
           <tr>
             <th></th>
             <th colspan="4">
-              <div [routerLink] = "['/addmentor']" class="ui right floated small primary labeled icon button">
+              <div [routerLink] = "['/dashboard/addmentor']" class="ui right floated small primary labeled icon button">
                 <i class="user icon"></i> Add Mentor
               </div>
               <div class="ui small  button">
