--- src/app/posts/posts-dashboard/posts-dashboard.component.scss
@@ -1,12 +1,0 @@
-section{
-    max-width: 700px;
-    margin: 2em auto;
-}
-
-mat-form-field{
-    width: 100%;
-}
-
-img{
-    width: 100%;
-}
