--- src/app/posts/posts-dashboard/posts-dashboard.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PostsDashboardComponent } from './posts-dashboard.component';
-
-describe('PostsDashboardComponent', () => {
-  let component: PostsDashboardComponent;
-  let fixture: ComponentFixture<PostsDashboardComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PostsDashboardComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(PostsDashboardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
