+++ src/app/student-user/student-user.component.html
@@ -74,7 +74,7 @@
   <tr>
     <th></th>
     <th colspan="4">
-      <div  [routerLink] = "['/adduser']" class="ui right floated small primary labeled icon button">
+      <div  [routerLink] = "['/dashboard/adduser']" class="ui right floated small primary labeled icon button">
         <i class="user icon"></i> Add User
       </div>
